+++ R/grouped-df.r
@@ -81,8 +81,13 @@ ungroup.grouped_df <- function(x, ...) {
 select_.grouped_df <- function(.data, ..., .dots) {
   dots <- lazyeval::all_dots(.dots, ...)
 
-  vars <- select_vars_(names(.data), dots,
-    include = as.character(groups(.data)))
+  group_vars <- as.character(groups(.data))
+  vars <- select_vars_(names(.data), dots, include = group_vars)
+
+  if (any(group_vars %in% vars)) {
+    warning("")
+  }
+
 
   select_impl(.data, vars)
 }